mavsdk_tests: split up Python script
Otherwise, the script will get a tangled mess.
